move assertions to their own module
